Moved CriticalMAAS command to subsystem and disabled by default
